requirements.txt
django==3.2.16
django-extensions==3.2.1
django-crispy-forms==1.14.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt[crypto]==5.2.2
djangorestframework-csv==2.1.1
drf-access-policy==1.3.0
